Updated Guardfile and assets systems to make Sass and compiled modular javascript optional

+++ lib/octopress/js_asset_manager.rb
@@ -10,25 +10,29 @@ module Octopress
     attr_reader :config
 
     def initialize
-      @js_assets_path = File.expand_path("../../assets/javascripts", File.dirname(__FILE__))
+      @js_assets_path = File.expand_path("../../javascripts", File.dirname(__FILE__))
 
-      unless Octopress.configuration.has_key? :require_js
-        abort "No :require_js key in configuration. Cannot proceed.".red
-      end
-      unless Octopress.configuration[:require_js].has_key? :lib
-        abort "No :lib key in :require_js configuration. Cannot proceed.".red
-      end
-      unless Octopress.configuration[:require_js].has_key? :modules
-        abort "No :modules key in :require_js configuration. Cannot proceed.".red
-      end
+      if Dir.exists? @js_assets_path
+        unless Octopress.configuration.has_key? :require_js
+          abort "No :require_js key in configuration. Cannot proceed.".red
+        end
+        unless Octopress.configuration[:require_js].has_key? :lib
+          abort "No :lib key in :require_js configuration. Cannot proceed.".red
+        end
+        unless Octopress.configuration[:require_js].has_key? :modules
+          abort "No :modules key in :require_js configuration. Cannot proceed.".red
+        end
 
-      # Read js dependencies from require_js.yml configuration
-      @lib = Octopress.configuration[:require_js][:lib].collect {|item| Dir.glob("#{@js_assets_path}/#{item}") }.flatten.uniq
-      @modules = Octopress.configuration[:require_js][:modules].collect {|item| "#{@js_assets_path}/#{item}" }.flatten.uniq
-      @module_files = @modules.collect {|item| Dir[item+'/**/*'] }.flatten.uniq
+        # Read js dependencies from require_js.yml configuration
+        @lib = Octopress.configuration[:require_js][:lib].collect {|item| Dir.glob("#{@js_assets_path}/#{item}") }.flatten.uniq
+        @modules = Octopress.configuration[:require_js][:modules].collect {|item| "#{@js_assets_path}/#{item}" }.flatten.uniq
+        @module_files = @modules.collect {|item| Dir[item+'/**/*'] }.flatten.uniq
 
-      @template_path = File.expand_path("../../#{Octopress.configuration[:source]}", File.dirname(__FILE__))
-      @build_path = "/javascripts/build"
+        @template_path = File.expand_path("../../#{Octopress.configuration[:source]}", File.dirname(__FILE__))
+        @build_path = "/javascripts/build"
+      else
+        @js_assets_path = false
+      end
     end
 
 
@@ -39,28 +43,36 @@ module Octopress
     end
 
     def url
-      Octopress.env == 'production' ?  "#{@build_path}/all-#{@fingerprint || get_fingerprint}.js" : "#{@build_path}/all.js"
+      if @js_assets_path
+        Octopress.env == 'production' ?  "#{@build_path}/all-#{@fingerprint || get_fingerprint}.js" : "#{@build_path}/all.js"
+      else
+        false
+      end
     end
 
     def compile
-      @fingerprint = get_fingerprint
+      if @js_assets_path
+        @fingerprint = get_fingerprint
 
-      filename = url
-      file = "#{@template_path + filename}"
+        filename = url
+        file = "#{@template_path + filename}"
 
-      if File.size?(file) && File.open(file) {|f| f.readline} =~ /#{@fingerprint}/
-        false
+        if File.size?(file) && File.open(file) {|f| f.readline} =~ /#{@fingerprint}/
+          false
+        else
+          js = Stitch::Package.new(:dependencies => @lib, :paths => @modules).compile
+          js = "/* Octopress fingerprint: #{@fingerprint} */\n" + js
+          js = Uglifier.new.compile js if Octopress.env == 'production'
+          write_path = "#{@template_path}/#{@build_path}"
+
+          (Dir["#{write_path}/*"]).each { |f| FileUtils.rm_rf(f) }
+          FileUtils.mkdir_p write_path
+          File.open(file, 'w') { |f| f.write js }
+
+          "Javascripts compiled to #{filename}."
+        end
       else
-        js = Stitch::Package.new(:dependencies => @lib, :paths => @modules).compile
-        js = "/* Octopress fingerprint: #{@fingerprint} */\n" + js
-        js = Uglifier.new.compile js if Octopress.env == 'production'
-        write_path = "#{@template_path}/#{@build_path}"
-
-        (Dir["#{write_path}/*"]).each { |f| FileUtils.rm_rf(f) }
-        FileUtils.mkdir_p write_path
-        File.open(file, 'w') { |f| f.write js }
-
-        "Javascripts compiled to #{filename}."
+        false
       end
     rescue Exception => e
       Octopress.logger.fatal "Error reading file #{url}".red
